Move health checks over to their own module
